Convert latitude and longitude to be a double everywhere in the code

+++ app/src/main/java/edu/usc/sunset/team7/www/parkhere/Services/URLEndpoints.java
@@ -13,7 +13,7 @@ import retrofit2.http.Path;
 
 public interface URLEndpoints {
     @GET("/search?lat={lat}?lon={lon}?startTime={startTime}?stopTime={stopTime}")
-    Call<SearchResult> getResults(@Path("lat") long lat, @Path("lon") long lon,
+    Call<SearchResult> getResults(@Path("lat") double lat, @Path("lon") double lon,
                                   @Path("startTime") long startTime, @Path("stopTime") long stopTime);
 
 }
